SCRUM-13 : Moved scheduler interval configuration to application.yml
