+++ src/main/resources/application.yml
@@ -14,6 +14,6 @@ spring:
   liquibase:
     change-log: classpath:db/changelog/master.yaml
 
-    scheduler:
-      interval: 3600000
+  scheduler:
+    interval: 3600000
